[refactor] pulls simulation out into separated class
